Compute Parent Adjacency Matrix with Multiple Approaches

Compute Parent Adjacency Matrix with Multiple Approaches

Usage

compute_parent_adjacency(
  ped,
  component,
  adjacency_method = "indexed",
  saveable,
  resume,
  save_path,
  verbose,
  lastComputed,
  nr,
  checkpoint_files,
  update_rate,
  parList,
  lens,
  save_rate_parlist,
  ...
)

Arguments

ped

a pedigree dataset. Needs ID, momID, and dadID columns

component

character. Which component of the pedigree to return. See Details.

adjacency_method

character. The method to use for computing the adjacency matrix. Options are "loop" or "indexed"

saveable

logical. If TRUE, save the intermediate results to disk

resume

logical. If TRUE, resume from a checkpoint

save_path

character. The path to save the checkpoint files

verbose

logical. If TRUE, print progress through stages of algorithm

lastComputed

the last computed index

nr

the number of rows in the pedigree dataset

checkpoint_files

a list of checkpoint files

update_rate

numeric. The rate at which to print progress

parList

a list of parent-child relationships

lens

a vector of the lengths of the parent-child relationships

save_rate_parlist

numeric. The rate at which to save the intermediate results by parent list. If NULL, defaults to save_rate*1000

...

additional arguments to be passed to ped2com

Details

The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions". For more advanced scenarios and detailed explanations, consult this vignette.