Summarize the paternal lines in a pedigree

Summarize the paternal lines in a pedigree

Usage

summarizePatrilines(
  ped,
  famID = "famID",
  personID = "ID",
  momID = "momID",
  dadID = "dadID",
  matID = "matID",
  patID = "patID",
  byr = NULL,
  founder_sort_var = NULL,
  include_founder = FALSE,
  nbiggest = 5,
  noldest = 5,
  skip_var = NULL,
  five_num_summary = FALSE,
  verbose = FALSE
)

Arguments

ped

a pedigree dataset. Needs ID, momID, and dadID columns

famID

character. Name of the column to be created in ped for the family ID variable

personID

character. Name of the column in ped for the person ID variable

momID

character. Name of the column in ped for the mother ID variable

dadID

character. Name of the column in ped for the father ID variable

matID

Character. Maternal line ID variable to be created and added to the pedigree

patID

Character. Paternal line ID variable to be created and added to the pedigree

byr

Optional column name for birth year.

founder_sort_var

The variable to sort the founders by. If NULL, the founders will be sorted by birth year (`byr`) if that's present and by `personID` otherwise.

include_founder

Logical, if TRUE, include the founder of each line in the summary statistics.

nbiggest

The number of biggest lines to return.

noldest

The number of oldest lines to return.

skip_var

A character vector of variables to skip when calculating summary statistics.

five_num_summary

Logical, if TRUE, include the 5-number summary (min, Q1, median, Q3, max) in the summary statistics.

verbose

Logical, if TRUE, print progress messages.

See also